Fixing the PHP. Also moving the gross mess of redirects to a new controller.
