+++ src/AppBundle/Controller/EventController.php
@@ -25,34 +25,35 @@ class EventController extends Controller
         $repo = $this->get('doctrine')
                      ->getManager()
                      ->getRepository('AppBundle\Entity\Event');
-        
-        
+
+
         $qb = $repo->createQueryBuilder('e');
         $qb
             ->where('e.date != :now')
             ->setParameter(':now', new \DateTime('yesterday'))
             ->orderBy('e.date', 'DESC');
         $events = $qb->getQuery()->getResult();
-        
+
         return $this->render('events/index.html.twig', [
             'events' => $events,
         ]);
     }
-    
+
     /**
-     * @Route("/{event_id}", name="event_view", requirements={"event_id": "\d+"})
+     * @Route("/{event_id}", name="event_view", requirements={
+     *    "event_id": "\d+"})
      */
     public function eventViewAction(Request $request, $event_id)
     {
         $em = $this->get('doctrine')->getManager();
         $event = $em->getRepository('AppBundle\Entity\Event')
                    ->find($event_id);
-        
+
         return $this->render('events/view.html.twig', [
             'event' => $event,
         ]);
     }
-    
+
     /**
      * @Route("/new", name="event_new")
      */
@@ -60,47 +61,47 @@ class EventController extends Controller
     {
         $event = new Event();
         $form = $this->createForm(EventType::class, $event);
-        
+
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
             // $file stores the uploaded image file
             $file = $event->getImage();
-            
+
             // Generate a unique name for the file before saving it
             $fileName = md5(uniqid()).'.'.$file->guessExtension();
-            
+
             // Move the file to the directory where brochures are stored
             $directory = $this->getParameter('event_image_directory');
             $file->move(
                 $directory,
                 $fileName
             );
-            
+
             // Create the full name from the upload directory
             // and the newly made filename
             $full_name = $directory . $fileName;
-            
+
             // Update the 'image' property to store the image file name
             // instead of its contents
             $event->setImage($full_name);
-            
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($event);
             $em->flush();
-            
+
             $event_id = $event->getId();
-            
+
             return $this->redirectToRoute('event_view', [
                 'event_id' => $event_id,
             ]);
         }
-        
+
         return $this->render('events/new.html.twig', [
             'form' => $form->createView(),
         ]);
     }
-    
+
     /**
      * @Route("/attendance", name="events_attendance")
      */
@@ -109,18 +110,19 @@ class EventController extends Controller
         $repo = $this->get('doctrine')
                      ->getManager()
                      ->getRepository('AppBundle\Entity\Event');
-        
+
         $qb = $repo->createQueryBuilder('e');
         $qb->orderBy('e.date', 'ASC');
         $events = $qb->getQuery()->getResult();
-        
+
         return $this->render('attendance/index.html.twig', [
             'events' => $events,
         ]);
     }
-    
+
     /**
-     * @Route("/{event_id}/attendance", name="event_attendance", requirements={"event_id": "\d+"})
+     * @Route("/{event_id}/attendance", name="event_attendance", requirements={
+     *    "event_id": "\d+"})
      * @Method({ "GET" })
      */
     public function eventAttendanceViewAction(Request $request, $event_id)
@@ -128,22 +130,23 @@ class EventController extends Controller
         $em = $this->get('doctrine')->getManager();
         $event = $em->getRepository('AppBundle\Entity\Event')
             ->find($event_id);
-        
+
         $query = $em->createQuery(
             'SELECT s.name, s.email
              FROM AppBundle\Entity\Attendance a
              JOIN a.event e
              JOIN a.attendee s
-             WHERE e.id = ?1 ');
+             WHERE e.id = ?1 '
+        );
         $query->setParameter(1, $event_id);
-        $attendees = $query->getResult();     
-        
+        $attendees = $query->getResult();
+
         return $this->render('attendance/event.html.twig', [
             'event' => $event,
             'attendees' => $attendees,
         ]);
     }
-    
+
     /**
      * @Route("/{event_id}/attendance", name="event_take_attendance", requirements={"event_id": "\d+"})
      * @Method({ "POST" })
@@ -154,16 +157,16 @@ class EventController extends Controller
         $em = $this->get('doctrine')->getManager();
         $event = $em->getRepository('AppBundle\Entity\Event')
             ->find($event_id);
-        
+
         // Now need to find all checked-in students
         $student_repo = $this->get('doctrine')
             ->getManager()
             ->getRepository('AppBundle\Entity\Student');
-        
+
         $qb = $student_repo->createQueryBuilder('s');
         $qb->where('s.checkedIn = true');
         $students = $qb->getQuery()->getResult();
-        
+
         // Iterate over the students and mark them as attendees
         foreach ($students as $student) {
             // Create the new attendance object
@@ -171,22 +174,22 @@ class EventController extends Controller
             $attendance->setAttendee($student);
             $attendance->setEvent($event);
             $event->addAttendee($attendance);
-            
+
             // Mark the student as not checked in
             $student->setCheckedIn(false);
-            
+
             // Update ALL the DB entries!!!
             $em->persist($attendance);
             $em->persist($student);
             $em->persist($event);
         }
         $em->flush();
-        
+
         return $this->redirectToRoute('event_attendance', [
             'event_id' => $event_id,
         ]);
     }
-    
+
     /**
      * Displays a form to edit an existing Event entity.
      *
@@ -198,34 +201,34 @@ class EventController extends Controller
         $em = $this->get('doctrine')->getManager();
         $event = $em->getRepository('AppBundle\Entity\Event')
                    ->find($event_id);
-        
+
         $prev_image_path = $event->getImage();
         $event->setImage(null);
-        
+
         $deleteForm = $this->createDeleteForm($event);
         $editForm = $this->createForm('AppBundle\Form\EventType', $event);
         $editForm->handleRequest($request);
 
         if ($editForm->isSubmitted() && $editForm->isValid()) {
-            
+
             // $file stores the uploaded image file
             $file = $event->getImage();
-            
+
             if (!empty($file)) {
                 // Generate a unique name for the file before saving it
                 $fileName = md5(uniqid()).'.'.$file->guessExtension();
-            
+
                 // Move the file to the directory where brochures are stored
                 $directory = $this->getParameter('event_image_directory');
                 $file->move(
                     $directory,
                     $fileName
                 );
-            
+
                 // Create the full name from the upload directory
                 // and the newly made filename
                 $full_name = $directory . $fileName;
-            
+
                 // Update the 'image' property to store the image file name
                 // instead of its contents
                 $event->setImage($full_name);
@@ -233,7 +236,7 @@ class EventController extends Controller
                 // They didn't upload a new image, so use the previous one
                 $event->setImage($prev_image_path);
             }
-            
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($event);
             $em->flush();
@@ -253,7 +256,8 @@ class EventController extends Controller
     /**
      * Deletes a Event entity.
      *
-     * @Route("/{event_id}", name="event_delete", requirements={"event_id": "\d+"})
+     * @Route("/{event_id}", name="event_delete", requirements={
+     *    "event_id": "\d+"})
      * @Method("DELETE")
      */
     public function eventDeleteAction(Request $request, $event_id)
@@ -261,7 +265,7 @@ class EventController extends Controller
         $em = $this->get('doctrine')->getManager();
         $event = $em->getRepository('AppBundle\Entity\Event')
                    ->findOneBy(array('id' => $event_id));
-        
+
         $form = $this->createDeleteForm($event);
         $form->handleRequest($request);
 
@@ -273,7 +277,7 @@ class EventController extends Controller
 
         return $this->redirectToRoute('events');
     }
-    
+
     /**
      * Creates a form to delete a Event entity.
      *
@@ -285,7 +289,7 @@ class EventController extends Controller
     {
         return $this->createFormBuilder()
             ->setAction($this->generateUrl('event_delete', [
-               'event_id' => $event->getId(), 
+               'event_id' => $event->getId(),
             ]))
             ->setMethod('DELETE')
             ->getForm()
