+++ src/AppBundle/Form/PostType.php
@@ -1,5 +1,5 @@
 <?php
-    
+
 namespace AppBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
