+++ src/AppBundle/Provider/UserProvider.php
@@ -11,50 +11,49 @@ use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 class UserProvider implements UserProviderInterface
 {
-  protected $userRepository;
+    protected $userRepository;
   
-  public function __construct(ObjectRepository $userRepository)
-  {
-      $this->userRepository = $userRepository;
-  }
+    public function __construct(ObjectRepository $userRepository)
+    {
+        $this->userRepository = $userRepository;
+    }
   
-  public function loadUserByUsername($username)
-  {
-      $q = $this->userRepository
+    public function loadUserByUsername($username)
+    {
+        $q = $this->userRepository
           ->createQueryBuilder('u')
           ->where('u.email = :email')
           ->setParameter('email', $username)
           ->getQuery();
-      try {
-          $user = $q->getSingleResult();
-      } catch (NoResultException $e) {
-          $message = sprintf(
+        try {
+            $user = $q->getSingleResult();
+        } catch (NoResultException $e) {
+            $message = sprintf(
               'Unable to find an active admin AppBundle:User object identified by "%s".',
               $username
           );
-          throw new UsernameNotFoundException($message, 0, $e);
-      }
-      return $user;
-  }
+            throw new UsernameNotFoundException($message, 0, $e);
+        }
+        return $user;
+    }
   
-  public function refreshUser(UserInterface $user)
-  {
-      $class = get_class($user);
-      if (!$this->supportsClass($class)) {
-          throw new UnsupportedUserException(
+    public function refreshUser(UserInterface $user)
+    {
+        $class = get_class($user);
+        if (!$this->supportsClass($class)) {
+            throw new UnsupportedUserException(
               sprintf(
                   'Instances of "%s" are not supported.',
                   $class
               )
           );
-      }
-      return $this->userRepository->find($user->getId());
-  }
+        }
+        return $this->userRepository->find($user->getId());
+    }
   
-  public function supportsClass($class)
-  {
-      return $this->userRepository->getClassName() === $class
+    public function supportsClass($class)
+    {
+        return $this->userRepository->getClassName() === $class
       || is_subclass_of($class, $this->userRepository->getClassName());
-  }
-  
+    }
 }
